+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Fix export of exception information in traces 
+([#1187](https://github.com/census-instrumentation/opencensus-python/pull/1187))
+
 ## 1.1.8
 
 Released 2023-01-18