+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.1.9
+
+Released 2023-03-10
+
 - Fix export of exception information in traces
 ([#1187](https://github.com/census-instrumentation/opencensus-python/pull/1187))
 - Modify metrics exporter to include setting export interval to 60s